third_party/blink/web_tests/fast/dom/shadow/host-pseudo-class-expected.txt

Test whether :host matches a shadow host correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

:host out of shadow tree should not match any shadow hosts.
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
:host with * should not match any shadow hosts.
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
:host with tag selector should not match any shadow hosts.
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
:host with class selector should not match any shadow hosts.
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
:host with id selector should not match any shadow hosts.
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
:host with attribute selector should not match any shadow hosts.
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
:host in a shadow tree should match its shadow host.
PASS backgroundColorOf('host') is "rgb(0, 128, 0)"
:host with :host-context in a shadow tree should match its shadow host.
PASS backgroundColorOf('host') is "rgb(0, 128, 0)"
:host takes a simple selector and matches when the simple selector matches.
PASS backgroundColorOf('host') is "rgb(0, 128, 0)"
:host matches a shadow host in just a nested shadow tree, not all enclosing shadow trees.
PASS backgroundColorOf('host1') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host1/host2') is "rgb(0, 128, 0)"
:host is updated when its matched host changes className or id.
PASS backgroundColorOf('host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
PASS backgroundColorOf('host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host2') is "rgba(0, 0, 0, 0)"
Compare :host with :host.
PASS backgroundColorOf('host') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host') is "rgb(255, 0, 0)"
PASS backgroundColorOf('host/target') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host/target') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host/target') is "rgb(0, 128, 0)"
PASS backgroundColorOf('host') is "rgba(0, 0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
